requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
image==1.5.33
Pillow==9.3.0
PyPDF2==2.11.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
tzdata==2022.6
